fix: use pointer for vector

+++ Project/point.cpp
@@ -1,6 +1,6 @@
 #include "point.h"
 
-Point::Point(double new_x, double new_y) : x(new_x), y(new_y)
+Point::Point(float new_x, float new_y) : x(new_x), y(new_y)
 {
 
 }
@@ -16,12 +16,12 @@ Point::Point() : Point(0, 0)
 
 }
 
-double Point::get_x()
+float Point::get_x()
 {
 	return x;
 }
 
-double Point::get_y()
+float Point::get_y()
 {
 	return y;
 }
@@ -36,17 +36,17 @@ int Point::int_y()
 	return (int) (y + 0.5);
 }
 
-double Point::dist(Point& that)
+float Point::dist(Point& that)
 {
 	return sqrt((x - that.x) * (x - that.x) + (y - that.y) * (y - that.y));
 }
 
-double Point::dir(Point& that)
+float Point::dir(Point& that)
 {
 	return atan2(that.y - y, that.x - x);
 }
 
-void Point::shift(double dx, double dy)
+void Point::shift(float dx, float dy)
 {
 	x += dx;
 	y += dy;
